% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{calculate.V_mean}
\alias{calculate.V_mean}
\title{Calculate the expected conditional variance term
This function calculates the expected conditional variance term Var(Xj |X-j)}
\usage{
calculate.V_mean(
  S,
  algo = "lasso",
  cv.rule = "min",
  out = NULL,
  Xmodel = "gaussian",
  sigma_X.list_S = NULL,
  verbose = FALSE
)
}
\arguments{
\item{S}{a list of selected variables.}

\item{algo}{a fitting algorithm (default: "lasso").}

\item{cv.rule}{indicates which rule should be used for the predict function, either "min" (the usual rule) or "1se" (the one-standard- error rule); default: "min").
See the glmnet help files for details.}

\item{out}{the fitted model from train.fun.}

\item{Xmodel}{model of the covaraites (default: "gaussian").}

\item{sigma_X.list_S}{a list of length |S|, with each element being the variance of the conditional distribution.}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
A vector of length |S|, whose element is the expected conditional variance term Var(Xj |X-j).
}
\description{
Calculate the expected conditional variance term
This function calculates the expected conditional variance term Var(Xj |X-j)
}
\details{
TBD
}
\seealso{
Other methods: 
\code{\link{calulate.mu_Xk}()},
\code{\link{fg.inference.binary}()},
\code{\link{fg.inference}()},
\code{\link{fit.mu}()},
\code{\link{floodgate.binary}()},
\code{\link{floodgate}()},
\code{\link{inference_general}()}
}
\concept{methods}
