% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{calulate.mu_Xk}
\alias{calulate.mu_Xk}
\title{Calculate mu(Xk) where Xk denotes the null sample
This function calculates mu(Xk) for null replicates Xk.}
\usage{
calulate.mu_Xk(
  binary = FALSE,
  X,
  i2,
  S,
  out,
  nulls.list_S = NULL,
  gamma_X.list_S = NULL,
  useMC = TRUE,
  Xmodel = "gaussian",
  algo = "lasso",
  predict.fun,
  cv.rule = "min",
  verbose = FALSE
)
}
\arguments{
\item{binary}{whether the response variable is binary (will be converted to factors if TRUE; default: FALSE).}

\item{X}{a n by p matrix, containing all the covariates.}

\item{i2}{the index of inference samples.}

\item{S}{a list of selected variables.}

\item{out}{the fitted model from train.fun.}

\item{nulls.list_S}{a list of length |S| whose element is a (|i2|*K)-dimensional vector, which contains
K set of null samples.}

\item{gamma_X.list_S}{a list of length |S|, with each element being the linear coefficient of
the given covariate on the other covariates (only relevant when Xmodel = "gaussian"; default: NULL).}

\item{useMC}{whether to use Monte Carlo estimators of the conditional quantities (default: TRUE).}

\item{Xmodel}{model of the covaraites (default: "gaussian").}

\item{algo}{a fitting algorithm (default: "lasso").}

\item{predict.fun}{a function to produce predictions of the response variable
with a given fitted model from train.fun and a matrix of new covariate values.}

\item{cv.rule}{indicates which rule should be used for the predict function, either "min" (the usual rule) or "1se" (the one-standard- error rule); default: "min").
See the glmnet help files for details.}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
A list of kength |S|, whose element is the matrix of mu(Xk) with dimension n2-by-K or n2-by-1.
}
\description{
Calculate mu(Xk) where Xk denotes the null sample
This function calculates mu(Xk) for null replicates Xk.
}
\details{
TBD
}
\references{
TBD
}
\seealso{
Other methods: 
\code{\link{calculate.V_mean}()},
\code{\link{fg.inference.binary}()},
\code{\link{fg.inference}()},
\code{\link{fit.mu}()},
\code{\link{floodgate.binary}()},
\code{\link{floodgate}()},
\code{\link{inference_general}()}
}
\concept{methods}
