% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.movi}
\alias{compute.movi}
\title{Compute the mMSE gap value}
\usage{
compute.movi(
  beta,
  Xmodel = "gaussian",
  Ydist = "gaussian",
  sigma_X.list = NULL,
  X = NULL,
  nulls.list = NULL
)
}
\arguments{
\item{beta}{a p-dimensional vector, true regression coefficients.}

\item{Xmodel}{model of the covaraites (default: "gaussian").}

\item{Ydist}{the type of conditional model.}

\item{sigma_X.list}{a list of length p, with each element being the variance of the conditional distribution.}

\item{X}{a n by p matrix, containing all the covariates.}

\item{nulls.list}{a list of length p, whose element is a (|i2|*K)-dimensional vector, which contains
K set of null samples.}
}
\value{
A vector of p, with each element being the value of I for a given covariate.
}
\description{
This function computes the mMSE gap value.
}
\details{
TBD
}
\references{
TBD
}
\concept{utils}
