% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{cosuff.g_copula.nulls}
\alias{cosuff.g_copula.nulls}
\title{Generate null samples for Gaussian copula covariates via co-sufficient sampling}
\usage{
cosuff.g_copula.nulls(X, i2, n21, S, K, sigma_X.list_S, verbose = FALSE)
}
\arguments{
\item{X}{a n by p matrix, containing all the covariates.}

\item{i2}{index of inference samples}

\item{n21}{number of batches, and the batch size will be |i2|/n21}

\item{S}{a subset of covariates, for which null samples are constructed.}

\item{K}{number of null replicates.}

\item{sigma_X.list_S}{a list of length |S|, with each element being the variance of the conditional distribution of the multivariate Gaussian.}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
nulls.list_S: a list of length |S| whose element is a (|i2|*K)-dimensional vector, which contains
K set of null samples.
}
\description{
This function generates co-sufficient null samples of a given subset of variables for Gaussian copula covariates.
}
\references{
\insertRef{LZ-LJ:2020}{floodgate}
}
\seealso{
Other sampling: 
\code{\link{cosuff.gaussian.nulls}()},
\code{\link{sample.g_copula.nulls}()},
\code{\link{sample.gaussian.nulls}()}
}
\concept{sampling}
