% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{cosuff.gaussian.nulls}
\alias{cosuff.gaussian.nulls}
\title{Generate null samples for multivariate Gaussian covariates via co-sufficient sampling}
\usage{
cosuff.gaussian.nulls(X, i2, n21, S, K, sigma_X.list_S, verbose = FALSE)
}
\arguments{
\item{X}{a n by p matrix, containing all the covariates.}

\item{i2}{index of inference samples}

\item{n21}{number of batches, and the batch size will be |i2|/n21}

\item{S}{a subset of covariates, for which null samples are constructed.}

\item{K}{number of null replicates.}

\item{sigma_X.list_S}{a list of length |S|, with each element being the variance of the conditional distribution.}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
A list of three objects.
nulls.list_S: a list of length |S| whose element is a (|i2|*K)-dimensional vector, which contains
K set of null samples;
cosuf.X_proj.list: a list of length |S| whose element is E(Xij|X-ij,Tm) for i in the m-th batch;
cosuf.sigma_X.list:a list of length |S| whose element is Var(Xij|X-ij,Tm).
}
\description{
This function generates co-sufficient null samples of a given subset of variables for multivariate Gaussian covariates.
}
\details{
TBD
}
\references{
TBD
}
\seealso{
Other sampling: 
\code{\link{cosuff.g_copula.nulls}()},
\code{\link{sample.g_copula.nulls}()},
\code{\link{sample.gaussian.nulls}()}
}
\concept{sampling}
