% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{fg.inference.binary}
\alias{fg.inference.binary}
\title{Core procedure of floodgate: binary responses}
\usage{
fg.inference.binary(
  S,
  mu_X,
  mu_Xk,
  Y,
  M_n = NULL,
  one.sided = TRUE,
  alevel = 0.05,
  test = "z",
  verbose = TRUE
)
}
\arguments{
\item{S}{a list of selected variables.}

\item{mu_X}{a list of kength |S|, whose element is the matrix of mu(X) with dimension n2-by-1.}

\item{mu_Xk}{a list of kength |S|, whose element is the matrix of mu(Xk) with dimension n2-by-K or n2-by-1.}

\item{Y}{a n2 by 1 matrix, containing the response variables of the inferene samples.}

\item{M_n}{the number of Monte Carlo samples for E(mu(X)|X-j).}

\item{one.sided}{whether to obtain LCB or p-values via the one-sided way (default: TRUE).}

\item{alevel}{confidence level (defaul: 0.05).}

\item{test}{type of the hypothesis test (defaul: "z").}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
A list of three objects.
inf.out: a matrix of |S|-by-4, containing the p-values, LCI, UCI and the floodgate LCB for variable in S;
S: a list of selected variables.
}
\description{
This function produces floodgate LCBs with given fitted mu (under binary responses).
}
\details{
TBD
}
\seealso{
Other methods: 
\code{\link{calculate.V_mean}()},
\code{\link{calulate.mu_Xk}()},
\code{\link{fg.inference}()},
\code{\link{fit.mu}()},
\code{\link{floodgate.binary}()},
\code{\link{floodgate}()},
\code{\link{inference_general}()}
}
\concept{methods}
