% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{fit.mu}
\alias{fit.mu}
\title{Fit a model on the training data for a given alogorithm.
This function fits a model on the training data for a given alogorithm.}
\usage{
fit.mu(
  X,
  Y,
  binary = FALSE,
  algo = "lasso",
  train.fun,
  active.fun,
  verbose = TRUE
)
}
\arguments{
\item{X}{a n by p matrix, containing all the covariates.}

\item{Y}{a n by 1 matrix, containing the response variables.}

\item{binary}{whether the response variable is binary (will be converted to factors if TRUE; default: FALSE).}

\item{algo}{a fitting algorithm.}

\item{train.fun}{a function to perform model fitting on the training data.}

\item{active.fun}{a function which takes the output of train.fun and outputs the selected variables.}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
A list of two objects. out: the fitted model from train.fun; S: a list of selected variables.
}
\description{
Fit a model on the training data for a given alogorithm.
This function fits a model on the training data for a given alogorithm.
}
\details{
TBD
}
\references{
TBD
}
\seealso{
Other methods: 
\code{\link{calculate.V_mean}()},
\code{\link{calulate.mu_Xk}()},
\code{\link{fg.inference.binary}()},
\code{\link{fg.inference}()},
\code{\link{floodgate.binary}()},
\code{\link{floodgate}()},
\code{\link{inference_general}()}
}
\concept{methods}
