% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{floodgate}
\alias{floodgate}
\title{Infer the model-free variable importance, mMSE gap via floodgate}
\usage{
floodgate(
  X,
  Y,
  i1,
  i2,
  nulls.list = NULL,
  gamma_X.list = NULL,
  sigma_X.list = NULL,
  Xmodel = "gaussian",
  funs,
  algo = "lasso",
  cv.rule = "min",
  one.sided = TRUE,
  alevel = 0.05,
  test = "z",
  verbose = TRUE
)
}
\arguments{
\item{X}{a n by p matrix, containing all the covariates.}

\item{Y}{a n by 1 matrix, containing the response variables.}

\item{i1}{the index of training samples.}

\item{i2}{the index of inference samples.}

\item{nulls.list}{a list of length p, whose element is a (|i2|*K)-dimensional vector, which contains
K set of null samples.}

\item{gamma_X.list}{a list of length p, with each element being the linear coefficient of
the given covariate on the other covariates (only relevant when Xmodel = "gaussian"; default: NULL).}

\item{sigma_X.list}{a list of length p, with each element being the variance of the conditional distribution.}

\item{Xmodel}{model of the covaraites (default: "gaussian").}

\item{funs}{a list of three: train.fun, active.fun and predict.fun.}

\item{algo}{a fitting algorithm (default: "lasso").}

\item{cv.rule}{indicates which rule should be used for the predict function, either "min" (the usual rule) or "1se" (the one-standard- error rule); default: "min").
See the glmnet help files for details.}

\item{one.sided}{whether to obtain LCB or p-values via the one-sided way (default: TRUE).}

\item{alevel}{confidence level (defaul: 0.05).}

\item{test}{type of the hypothesis test (defaul: "z").}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
A list of three objects.
inf.out: a matrix of |S|-by-4, containing the p-values, LCI, UCI and the floodgate LCB for variable in S;
S: a list of selected variables;
cpu.time: computing time.
}
\description{
This function infers the model-free variable importance, mMSE gap via floodgate.
}
\references{
\insertRef{LZ-LJ:2020}{floodgate}
}
\seealso{
Other methods: 
\code{\link{calculate.V_mean}()},
\code{\link{calulate.mu_Xk}()},
\code{\link{fg.inference.binary}()},
\code{\link{fg.inference}()},
\code{\link{fit.mu}()},
\code{\link{floodgate.binary}()},
\code{\link{inference_general}()}
}
\concept{methods}
