% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\docType{data}
\name{funs.list}
\alias{funs.list}
\title{A list of functions.}
\format{
An object of class \code{list} of length 6.
}
\usage{
funs.list
}
\description{
A list of functions, which specify various fitting algorithms,
including lasso, step, ridge,
binom_lasso, binom_ridge, pois,
rf, rf_binary.
}
\keyword{datasets}
