% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{Genotypes.parm}
\alias{Genotypes.parm}
\title{Obtain parameters from a fastPHASE genotype model}
\usage{
Genotypes.parm(p)
}
\arguments{
\item{p}{number of nodes}
}
\value{
A list containing pInit: a vector of K_st, which gives the marginal probabilities of the first variable;
Q: an array of size (p-1, K_st, K_st), which gives the p-1 transition matrices of the hidden Markov chain;
pEmit: an array of size (p, M, K_st), which is the emission probability matrix.
}
\description{
This function obtains the HMM parameters from a fastPHASE genotype model
}
\details{
The first p variables also follow a HMM.
}
\seealso{
Other models: 
\code{\link{HMM.parm}()},
\code{\link{gen.Y}()}
}
\concept{models}
