% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{HMM.parm}
\alias{HMM.parm}
\title{Obtain model parameters from a randomly generated HMM}
\usage{
HMM.parm(p, K_st, M)
}
\arguments{
\item{p}{number of nodes}

\item{K_st}{number of hidden states}

\item{M}{number of emission states}
}
\value{
A list containing pInit: a vector of K_st, which gives the marginal probabilities of the first variable;
Q: an array of size (p-1, K_st, K_st), which gives the p-1 transition matrices of the hidden Markov chain;
pEmit: an array of size (p, M, K_st), which is the emission probability matrix.
}
\description{
This function obtains model parameters from a randomly generated hidden Markov model
}
\references{
\insertRef{sesia2019}{floodgate}
}
\seealso{
Other models: 
\code{\link{Genotypes.parm}()},
\code{\link{gen.Y}()}
}
\concept{models}
