% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{inference_general}
\alias{inference_general}
\title{Obtain the floodgate LCB using the Delta method}
\usage{
inference_general(
  R,
  V = NULL,
  V_mean = NULL,
  alevel = 0.05,
  test = "z",
  one.sided = TRUE
)
}
\arguments{
\item{R}{a vector of length n2, corresonding the term in the paper.}

\item{V}{a vector of length n2, corresonding the term in the paper (defaul: NULL).}

\item{V_mean}{the expected conditional variance term Var(Xj |X-j) (defaul: NULL).}

\item{alevel}{confidence level (defaul: 0.05).}

\item{test}{type of the hypothesis test (defaul: "z").}

\item{one.sided}{whether to obtain LCB or p-values via the one-sided way (default: TRUE).}
}
\value{
A vector of length 4: p-value, LCI, UCI and LCB.
}
\description{
This function produces floodgate LCB using the Delta method.
}
\references{
\insertRef{LZ-LJ:2020}{floodgate}
}
\seealso{
Other methods: 
\code{\link{calculate.V_mean}()},
\code{\link{calulate.mu_Xk}()},
\code{\link{fg.inference.binary}()},
\code{\link{fg.inference}()},
\code{\link{fit.mu}()},
\code{\link{floodgate.binary}()},
\code{\link{floodgate}()}
}
\concept{methods}
