% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample.g_copula.nulls}
\alias{sample.g_copula.nulls}
\title{Generate null samples for Gaussian copula covariates
This function generates null samples of a given subset of variables for Gaussian copula covariates.}
\usage{
sample.g_copula.nulls(X, S, K, gamma_X.list_S, sigma_X.list_S, verbose = FALSE)
}
\arguments{
\item{X}{a n by p matrix, containing the Gaussian vectors before applying the probability integral transform}

\item{S}{a subset of covariates, for which null samples are constructed.}

\item{K}{number of null replicates.}

\item{gamma_X.list_S}{a list of length |S|, with each element being the linear coefficient of
the given covariate on the other covariates (before the probability integral transform).}

\item{sigma_X.list_S}{a list of length |S|, with each element being the variance of the conditional distribution
(before the probability integral transform).}

\item{verbose}{whether to show intermediate progress (default: FALSE).}
}
\value{
A list of length |S|, with each element being a (n*K)-dimensional vector, which contains
K set of null samples.
}
\description{
Generate null samples for Gaussian copula covariates
This function generates null samples of a given subset of variables for Gaussian copula covariates.
}
\details{
TBD
}
\references{
TBD
}
\seealso{
Other sampling: 
\code{\link{cosuff.g_copula.nulls}()},
\code{\link{cosuff.gaussian.nulls}()},
\code{\link{sample.gaussian.nulls}()}
}
\concept{sampling}
